Extract shared, pan-related code from FreeScrollState and PanZoomState - New BasePanState abstract class in charge of pan movement.
